% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convert_inspire_ids_rcpp}
\alias{convert_inspire_ids_rcpp}
\title{Convert INSPIRE IDs between long and short formats (C++ Kernel)}
\usage{
convert_inspire_ids_rcpp(ids, crs, axis_order)
}
\arguments{
\item{ids}{Character vector of IDs to convert.}

\item{crs}{Integer EPSG code (used when converting short to long).}

\item{axis_order}{String "NE" or "EN" (used when converting long to short).}
}
\value{
Character vector of converted IDs.
}
\description{
Convert INSPIRE IDs between long and short formats (C++ Kernel)
}
\keyword{internal}
