% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_grid.R
\name{derive_grid}
\alias{derive_grid}
\title{Convert INSPIRE IDs into a spatial grid}
\usage{
derive_grid(
  ids,
  point_type = c("llc", "centroid"),
  output_type = c("sf_polygons", "sf_points", "dataframe"),
  include_llc = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{ids}{A vector containing character strings of INSPIRE-compliant IDs.
Can be either short or long INSPIRE IDs.}

\item{point_type}{A character string, used only when \code{output_type = "sf_points"}.
Determines the location of the points: \code{"centroid"} for the center
of the cell, or \code{"llc"} (default) for the lower-left corner.}

\item{output_type}{The class of the output object: \code{"sf_polygons"} (default) creates
a spatial object with polygon geometries, \code{"sf_points"} creates an \code{sf}
object with point geometries, and \code{"dataframe"} creates a data frame with
grid cell centroid coordinates (\code{X_centroid}, \code{Y_centroid}).}

\item{include_llc}{A logical value. If \code{TRUE} (default), columns for the
lower-left corner coordinates (\code{X_LLC}, \code{Y_LLC}) of each cell are included
in the output.}

\item{quiet}{Logical value. If \code{TRUE}, all progress messages are suppressed.
Defaults to \code{FALSE}.}
}
\value{
An \code{sf} object or \code{data.frame} representing the grid derived from
the INSPIRE IDs.
}
\description{
{
This function takes a vector of INSPIRE-compliant IDs and derives a
regular spatial grid from it. For generating a spatial grid from a spatial
extent, see \code{\link{create_grid}}.}
}
\examples{
library(sf)

inspire <- c(
  "CRS3035RES100000mN26E43", "CRS3035RES100000mN26E44",
  "CRS3035RES100000mN27E41", "CRS3035RES100000mN27E42",
  "CRS3035RES100000mN27E43", "CRS3035RES100000mN27E44"
)

grid <- derive_grid(inspire)
plot(grid$geometry)
}
