% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generate_ids_rcpp}
\alias{generate_ids_rcpp}
\title{Generate INSPIRE IDs Only (C++ Kernel)}
\usage{
generate_ids_rcpp(
  x_llc,
  y_llc,
  cellsize,
  epsg,
  size_lbl,
  divider,
  axis_order,
  id_format
)
}
\arguments{
\item{x_llc}{Numeric vector of X coordinates (bottom-left)}

\item{y_llc}{Numeric vector of Y coordinates (bottom-left)}

\item{cellsize}{Grid resolution in meters}

\item{epsg}{EPSG code}

\item{size_lbl}{Label for short ID}

\item{divider}{Divisor for short ID coordinates}

\item{axis_order}{"NE" or "EN"}

\item{id_format}{"both", "long", "short"}
}
\value{
List containing IDs
}
\description{
Efficiently generates INSPIRE IDs for a set of coordinates.
}
