% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridmaker-package.R
\docType{package}
\name{gridmaker-package}
\alias{gridmaker}
\alias{gridmaker-package}
\title{gridmaker: Create INSPIRE-Compliant Grids with IDs}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Creates GISCO compatible and INSPIRE-compliant grids with IDs that look like 'CRS3035RES1000mN3497000E4448000' or '1kmN3497E4447'. Input can be 'sf', 'sfc' objects or bounding boxes. Output can be 'sf' polygons, 'sf' centroids, or just 'data.frame' with grid cell center or bottom left corner coordinates. The resulting grids are always aligned to rounded coordinates as per INSPIRE requirements.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/e-kotov/gridmaker}
  \item \url{https://www.ekotov.pro/gridmaker/}
  \item Report bugs at \url{https://github.com/e-kotov/gridmaker/issues}
}

}
\author{
\strong{Maintainer}: Egor Kotov \email{kotov.egor@gmail.com} (\href{https://orcid.org/0000-0001-6690-5345}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Jonas Lieth \email{jslth@outlook.com} (\href{https://orcid.org/0000-0002-3451-3176}{ORCID}) [contributor]
}

}
\keyword{internal}
