% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspire_convert.R, R/inspire_extract.R,
%   R/inspire_generate.R
\name{inspire}
\alias{inspire}
\alias{inspire_convert}
\alias{inspire_extract}
\alias{inspire_generate}
\title{Convert INSPIRE IDs between long and short formats}
\usage{
inspire_convert(ids, crs = 3035)

inspire_extract(inspire, as_sf = FALSE, crs = NULL)

inspire_generate(
  coords,
  cellsize_m = NULL,
  short = FALSE,
  llc = FALSE,
  tolerance = 1e-06,
  sample = 2000
)
}
\arguments{
\item{ids}{A character vector of INSPIRE IDs. All IDs in the vector must be
of the same format (either all long or all short).}

\item{crs}{An optional numeric EPSG code or \code{sf::st_crs} object. If provided,
this CRS will be assigned to all parsed coordinates, overriding any CRS
information found in long-form IDs. If \code{NULL} (the default), the CRS is
inferred from long-form IDs. When only short-form IDs are present,
the function will default to EPSG:3035 (with a warning) for both
\code{sf} and \code{data.frame} outputs.}

\item{inspire}{A vector of INSPIRE IDs. Can be either legacy or non-legacy.}

\item{as_sf}{Whether to return an object of class \code{sfc} or a dataframe.}

\item{coords}{A list, matrix, or dataframe where the X and Y coordinates are
either in the columns \code{"x"} and \code{"y"} or in the first and second
column position, respectively. Column names are converted to lowercase.

Can also be a \code{sf}/\code{sfc} object in which case the coordinates are
extracted using \code{\link[sf]{st_coordinates}}.}

\item{cellsize_m}{A single integer representing the grid cell size in metres
(e.g., 1000 for a 1 km grid).}

\item{short}{If \code{TRUE}, generates short INSPIRE ID. Defaults to
\code{FALSE}.}

\item{llc}{Do the coordinates in \code{coords} represent the lower-left
corners of their cells? If \code{FALSE}, subtracts each coordinate by
half of \code{cellsize_m}. If \code{TRUE}, leaves them as-is. Defaults
to \code{FALSE}, i.e., treat coordinates as cell centroids.}

\item{tolerance}{If \code{res} is \code{NULL}, controls the maximum
acceptable difference between calculated cell spacings to consider them
uniform. Defaults to \code{1e-6}.}

\item{sample}{If \code{res} is \code{NULL}, specifies the number of points
to guess a resolution from. Defaults to 2000 to keep performance high.
Increase this value if you are uncertain about the quality of your data.}
}
\value{
A character vector of the converted INSPIRE IDs.

\code{z22_inspire_generate} returns a character vector containing
the INSPIRE identifiers.

\code{z22_inspire_extract} returns a dataframe or
\code{\link[sf:st_sf]{sf}} dataframe (if \code{as_sf = TRUE}) containing
the points extracted from the INSPIRE identifiers and information about
the CRS and cell sizes. Note that the returned coordinates are always
the centers of the grid cells as opposed to the lower-left corners.
}
\description{
This function converts a vector of INSPIRE-compliant grid IDs from their long
format to the short format, or vice-versa. It automatically detects the
input format and is fully vectorized to handle large inputs efficiently.

The long format is structured as \verb{CRS\{epsg\}RES\{cellsize\}mN\{y\}E\{x\}}, while the
short format is \verb{\{cellsize\}N\{y\}E\{x\}}.

Given pairs of coordinates, generates their INSPIRE grid representation.
Given INSPIRE identifiers, can also extract the X and Y coordinates.

An INSPIRE ID contains information about the CRS, cell size and the
ETRS89-LAEA coordinates of the lower-left corner of the grid cell in its
format.

\preformatted{CRS3035{cellsize}mN{y}E{x} # long format
{cellsize}N{y}E{x}         # short format}

The long format always uses meters while the short format aggregates
cell sizes greater or equal to 1000m to km.
}
\examples{
long_ids <- c("CRS3035RES1000mN2684000E4334000", "CRS3035RES10000mN2700000E4400000")
short_ids <- c("1kmN2684E4334", "10kmN270E440")

# Convert long to short
inspire_convert(long_ids)

# Convert short to long
inspire_convert(short_ids)

# Convert short to long with a different CRS
inspire_convert(short_ids, crs = 3857)
# Generate IDs from a dataframe
coords <- data.frame(x = c(4334100, 4334200), y = 2684000)
gen <- inspire_generate(coords, llc = TRUE)
ext <- inspire_extract(gen)[c("x", "y")]
identical(ext, coords)

# Extract coordinates from short ID strings
inspire_extract("100mN34000E44000")

# Generate IDs from an sf dataframe
if (requireNamespace("sf", quietly = TRUE)) {
  coords <- sf::st_as_sf(coords, coords = c("x", "y"))
  inspire_generate(coords)
}
}
