% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parse_inspire_ids_rcpp}
\alias{parse_inspire_ids_rcpp}
\title{Parse INSPIRE IDs to Coordinates (C++ Kernel)}
\usage{
parse_inspire_ids_rcpp(inspire, is_long, is_short)
}
\arguments{
\item{inspire}{Character vector of IDs}

\item{is_long}{Logical indicator for long IDs}

\item{is_short}{Logical indicator for short IDs}
}
\value{
DataFrame with columns: crs, cellsize, x, y
}
\description{
Parse INSPIRE IDs to Coordinates (C++ Kernel)
}
\keyword{internal}
