% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridMaker-package.R
\docType{package}
\name{GridMaker-package}
\alias{GridMaker-package}
\title{GridMaker: Vector GIS grid datasets}
\description{
GridMaker produces GIS grid datasets of various resolutions covering user-defined countries or regions. An R-translation of the Eurostat-GISCO GridMaker java utility.
}
\author{
\strong{Maintainer}: Francisco J. Goerlich \email{Francisco.J.Goerlich@uv.es} (\href{https://orcid.org/0000-0003-1626-525X}{ORCID})

}
\keyword{internal}
